Move stress worker to new file
